refactor(convertion): move linear-system pattern out of generic helpers
