+++ src/data/helpers/index.ts
@@ -1,6 +1,5 @@
 export * from './wrappers';
 export * from './join-with-many-separators';
-export * from './linear-system-pattern';
 export * from './mathml-element-to-latex-converter';
 export * from './normalize-whitespace';
 export * from './own-lookup';
